Create empty stages, move stages 2,3 to 7,8
