+++ BowGame/StageFive.swift
@@ -1,0 +1,13 @@
+//
+//  StageFive.swift
+//  BowGame
+//
+//  Created by Zhiyang Lu on 11/10/15.
+//  Copyright © 2015 Antonis papantoniou. All rights reserved.
+//
+
+import UIKit
+
+class StageFive: StageGameScene {
+
+}
