+++ BowGame/StageFour.swift
@@ -1,0 +1,13 @@
+//
+//  StageFour.swift
+//  BowGame
+//
+//  Created by Zhiyang Lu on 11/10/15.
+//  Copyright © 2015 Antonis papantoniou. All rights reserved.
+//
+
+import UIKit
+
+class StageFour: StageGameScene {
+
+}
