+++ BowGame/StageSix.swift
@@ -1,0 +1,14 @@
+
+//
+//  StageSix.swift
+//  BowGame
+//
+//  Created by Zhiyang Lu on 11/10/15.
+//  Copyright © 2015 Antonis papantoniou. All rights reserved.
+//
+
+import UIKit
+
+class StageSix: StageGameScene {
+
+}
